Add static cast

+++ image_generation/stable_diffusion_1_5/cpp/src/main.cpp
@@ -398,7 +398,7 @@ int32_t main(int32_t argc, char* argv[]) try {
         std::uint32_t seed = num_images == 1 ? user_seed : user_seed + n;
 
         // latents are multiplied by 'init_noise_sigma'
-        ov::Shape latent_shape = ov::Shape({batch_size, sample_shape[1].get_length(), height / VAE_SCALE_FACTOR, width / VAE_SCALE_FACTOR});
+        ov::Shape latent_shape = ov::Shape({batch_size, static_cast<size_t>(sample_shape[1].get_length()), height / VAE_SCALE_FACTOR, width / VAE_SCALE_FACTOR});
         ov::Shape latent_model_input_shape = latent_shape;
         ov::Tensor noise = randn_tensor(latent_shape, read_np_latent, seed);
         latent_model_input_shape[0] = 2;  // Unet accepts batch 2
